api/microscope/requirements.txt
aioice==0.9.0
aiortc==1.5.0
anyio==3.7.1
argcomplete==3.1.2
asyncio==3.4.3
av==10.0.0
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
commitizen==3.12.0
cryptography==41.0.4
decli==0.6.1
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.103.2
google-crc32c==1.5.0
gpio==1.0.0
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
passlib==1.7.4
prompt-toolkit==3.0.36
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.13
pyee==11.0.1
pylibsrtp==0.8.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
questionary==2.0.1
RPi.GPIO==0.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
starlette==0.27.0
termcolor==2.3.0
tomlkit==0.12.1
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
wcwidth==0.2.8
websockets==11.0.3
zipp==3.17.0
